Removed A3C since the only one that will ever be trained and evaluated is A2C.
